Extra credits tasks implemnted
Additional placeholder text for other text fields.
Use the required attribute to add HTML5 validation to make sure that
required fields are filled out and input is formatted correctly.
Add at least the following additional styling enhancements to the form,
feel free to extra styling but leave the basic layout the same as the
mockup.:
Change the background color for at least ONE of the main sections of
the site.
Uses CSS transitions for focus states.
Use CSS transition effects for highlighting the look of the form fields.
For example, make a background color fade into view when the field the
user clicks into a text field, and fade out when the user clicks or
tabs out of the field.